Add tests for allowed/unallowed merges

+++ Tests/post-merge.Tests.ps1
@@ -21,7 +21,7 @@ Import-Module $poshUnitModuleFile
 . "$PSScriptRoot\TestHelpers.ps1"
 . "$PSScriptRoot\..\Tools\GitHooks\Common.ps1"
 
-Test-Fixture "post-merge hooks tests" `
+Test-Fixture "post-merge hooks tests for non-conflict pull merge" `
     -SetUp `
     {
         $tempPath = Get-TempTestPath
@@ -59,7 +59,7 @@ Test-Fixture "post-merge hooks tests" `
         {
             Pop-Location
 
-            if (-not $externalProcess.HasExited)
+            if (($externalProcess -ne $null) -and (-not $externalProcess.HasExited))
             {
                 taskkill /PID $($externalProcess.Id) /F /T
             }
@@ -141,5 +141,72 @@ Test-Fixture "post-merge hooks tests" `
 
             $setting = git config branch.master.rebase
             $Assert::That($setting, $Is::EqualTo("true"))
+        }
+    )
+
+Test-Fixture "post-merge hooks tests for allowed and unallowed non-conflict merges" `
+    -SetUp `
+    {
+        $tempPath = Get-TempTestPath
+        $localRepoPath = Prepare-LocalGitRepo $tempPath
+
+        $remoteRepoPath = "$tempPath\RemoteGitRepo"
+        New-Item -Path $remoteRepoPath -ItemType Directory
+        Push-Location $remoteRepoPath
+        git init --bare
+        Pop-Location
+
+        Push-Location $localRepoPath
+        git remote add origin $remoteRepoPath
+        git push origin master --set-upstream
+        tools\GitHooks\Install-GitHooks.ps1 post-merge
+
+        New-Item -Path "ReadyForRelease10.txt" -ItemType File
+        git add "ReadyForRelease10.txt"
+        git commit -m "Ready for release 1.0"
+        git push origin master
+        git checkout -b release.1.0
+        New-Item -Path "FixForRelease10.txt" -ItemType File
+        git add "FixForRelease10.txt"
+        git commit -m "Fix for release 1.0"
+        git push origin release.1.0 --set-upstream
+        git checkout master
+        New-Item -Path "FeatureForFutureReleases.txt" -ItemType File
+        git add "FeatureForFutureReleases.txt"
+        git commit -m "Fix for future releases"
+
+        $externalProcess = $null
+    } `
+    -TearDown `
+    {
+        Pop-Location
+
+        if (($externalProcess -ne $null) -and (-not $externalProcess.HasExited))
+        {
+            taskkill /PID $($externalProcess.Id) /F /T
+        }
+
+        Remove-Item -Path $tempPath -Recurse -Force
+    } `
+    -Tests `
+    (
+        Test "Merge allowed branches from configuration is made as is" `
+        {
+            git merge release.1.0
+
+            $Assert::That((Test-MergeCommit), $Is::True)
+        }
+    ),
+    (
+        Test "Merge unallowed branches from configuration prompts UI dialog" `
+        {
+            git checkout release.1.0
+            $externalProcess = Start-PowerShell { git merge master }
+
+            Init-UIAutomation
+
+            $dialog = Get-UIAWindow -Name "Unallowed merge"
+
+            $Assert::That($dialog, $Is::Not.Null)
         }
     )
